Log receivers of event.
